gltf: Move shared structures into gltfpack.h
This file will be included by all sources; it's still a monolith but
it's probably fine to keep it as is for the time being.
